Add files via upload
I wrote the code and then I checked the console but it said there was a problem with line 3 in p5.min.js even though I am not using p5.min.js. It also said that there was a problem with my background and my rocket image for some reason.